% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mandrill_send_template}
\alias{mandrill_send_template}
\title{Send an email with template}
\usage{
mandrill_send_template(api_key, template, content, email)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{An optional list which can contain an amount to refund}
}
\value{
nothing
}
\description{
This sends an email with template
}

