% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_utilities.R
\name{process_attachements}
\alias{process_attachements}
\title{Encode attachment to base64 and provide meta data for JSON payload}
\usage{
process_attachements(attachment)
}
\arguments{
\item{attachment}{character vector of unencoded attachment file location}
}
\value{
data frame with attachement type, attachment name (which matches CID in Mandrill template), and
base 64 encoded attachement.
}
\description{
Encode attachment to base64 and provide meta data for JSON payload
}

