% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_utilities.R
\name{process_contents}
\alias{process_contents}
\title{Process HTML contents passed to `mc:edit` classes in Mandrill template}
\usage{
process_contents(content)
}
\arguments{
\item{character}{vector of HTML to replace contents btween HTML tags in Mandrill Template}
}
\value{
data frame with content name (matching `mc:edit="content_name"` in template)
}
\description{
Process HTML contents passed to `mc:edit` classes in Mandrill template
}

