% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_utilities.R
\name{process_images}
\alias{process_images}
\title{Encode PNG to base64 and provide meta data for JSON payload}
\usage{
process_images(image)
}
\arguments{
\item{image}{character vector of unencoded PNG file location}
}
\value{
data frame with image type, image name (which matches CID in Mandrill template), and
base 64 encoded image.
}
\description{
Encode PNG to base64 and provide meta data for JSON payload
}

