% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/send.R
\name{mandrill_send_template}
\alias{mandrill_send_template}
\title{Send an email with template}
\usage{
mandrill_send_template(api_key = NA, template_name = NA, variables = NA,
  subject = NA, recipient = NA, sender = NA, contents = NULL,
  images = NULL, css = FALSE)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{template_name}{name of template on Mandrill to use}

\item{variables}{list of variables used in template to be filled in}

\item{subject}{character vector (single element) to populate email subject line}

\item{recipient}{character vector of email addresses of recipients}

\item{sender}{characer vector of sendes' email addresses}

\item{contents}{contents of email}

\item{images}{character vector of PNGs to include in email.  names (minus the suffix)
must match the image declaration in the Mandrill template. Also note that images must
saved in same directory as calling environment.}

\item{css}{whether to include inline CSS or not.}
}
\value{
nothing
}
\description{
This sends an email with template
}

