% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/send.R
\name{mandrill_send_template}
\alias{mandrill_send_template}
\title{Send an email with template}
\usage{
mandrill_send_template(api_key = NA, template_name = NA,
  global_variables = NA, subject = NA, recipient = NA, sender = NA,
  contents = NULL, images = NULL, css = FALSE)
}
\arguments{
\item{api_key}{your Mandrill API Key.}

\item{template_name}{name of template on Mandrill to use}

\item{global_variables}{list of global merge variables used in template to be filled in.
Mandrill expects these to come as an array, with each key/value pair split into
name: key, content: value.  for instance, if your key was named 'orderid', and the value was
'order1019', Mandrill would expect the JSON for these merge variables to look like

\code{`global_merge_vars": [{"name": "orderid", "content": "order1019"}]`}.

one easy way to pass these variables is as a data frame.  using our example above,

\code{global_variables = data.frame('name'=c('orderid'), 'content'=c('order1019'))}

will be converted by jsonlite::toJSON into an JSON array that Mandrill can understand.}

\item{subject}{character vector (single element) to populate email subject line}

\item{recipient}{character vector of email addresses of recipients, or data.frame with
headers "email", "name", and "type", per mandrill documentation.}

\item{sender}{character vector of senders' email addresses}

\item{contents}{list HTML contents to be replaced in Mandrill template.  These are used to replace content
between HTML tags that have a CSS class of `mc:edit="content_name"` where the name of the object in the list
must match the the `content_name` in in the template.}

\item{images}{character vector denoting location PNGs to include in email. PNG file names (minus the suffix and
any preceeding path directories) must match the image declaration in the Mandrill template.}

\item{css}{whether to include inline CSS or not.  This CSS must be already present in the Mandrill template}
}
\value{
contents of HTTP Response.
}
\description{
This sends an email with template
}
\references{
\url{https://mandrillapp.com/api/docs/messages.JSON.html#method=send} has examples
of valid JSON for sending a message using Mandrill.
}

