% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/send.R
\name{mandrill_send_template}
\alias{mandrill_send_template}
\title{Send an email with template}
\usage{
mandrill_send_template(api_key = NA, template_name = NA, variables = NA,
  subject = NA, recipient = NA, sender = NA, contents = NULL,
  images = NULL, css = FALSE)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{template_name}{name of template on Mandrill to use}

\item{variables}{list of merge variables used in template to be filled in}

\item{subject}{character vector (single element) to populate email subject line}

\item{recipient}{character vector of email addresses of recipients}

\item{sender}{character vector of senders' email addresses}

\item{contents}{list HTML  contents to be replaced in Mandrill template.  These are used to replace content
between HTML tags that have a CSS class of `mc:edit="content_name"` where the name of the object in the list
must match the the `content_name` in in the template.}

\item{images}{character vector denoting location PNGs to include in email. PNG file names (minus the suffix and
any preceeding path directories) must match the image declaration in the Mandrill template.}

\item{css}{whether to include inline CSS or not.  This CSS must be already present in the Mandrill template}
}
\value{
contents of HTTP Response.
}
\description{
This sends an email with template
}

